finishes decoding
