+++ libs/moloch-v3-macro-ui/src/components/ProposalDetails/ProposalDetailsContainer.tsx
@@ -68,7 +68,6 @@ export const ProposalDetailsContainer = ({
         actionData,
       });
 
-      console.log('proposalActions', proposalActions);
       if (shouldUpdate) {
         setActionData(proposalActions);
         setDecodeError(
